refactor(front): Renamed `Logs` to `LogsPreview`
