+++ front/components/LogsPreview.vue
